Move CoreGroup into a new directory (#637)
