+++ openmls/src/group/core_group/test_duplicate_extension.rs
@@ -2,7 +2,7 @@
 use super::*;
 
 use crate::{messages::GroupSecrets, prelude::*, schedule::KeySchedule, test_utils::*};
-use mls_group::create_commit_params::CreateCommitParams;
+use core_group::create_commit_params::CreateCommitParams;
 use openmls_rust_crypto::OpenMlsRustCrypto;
 use openmls_traits::crypto::OpenMlsCrypto;
 use openmls_traits::OpenMlsCryptoProvider;
