+++ openmls/src/group/managed_group/membership.rs
@@ -1,7 +1,7 @@
 #[cfg(any(feature = "test-utils", test))]
 use std::collections::BTreeMap;
 
-use mls_group::create_commit_params::CreateCommitParams;
+use core_group::create_commit_params::CreateCommitParams;
 
 use super::*;
