+++ openmls/src/schedule/mod.rs
@@ -623,7 +623,7 @@ impl ExporterSecret {
 
     /// Derive a `Secret` from the exporter secret. We return `Vec<u8>` here, so
     /// it can be used outside of OpenMLS. This function is made available for
-    /// use from the outside through [`crate::group::mls_group::export_secret`].
+    /// use from the outside through [`crate::group::core_group::export_secret`].
     pub(crate) fn derive_exported_secret(
         &self,
         ciphersuite: &Ciphersuite,
